Re-started integrating the new-new graphics handler. #54
